+++ src/test/java/sparta/todoapp/domain/auth/controller/AuthControllerTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
-import sparta.todoapp.domain.auth.dto.AuthLoginRequestDto;
-import sparta.todoapp.domain.auth.dto.AuthSignUpRequestDto;
+import sparta.todoapp.domain.auth.dto.request.AuthLoginRequestDto;
+import sparta.todoapp.domain.auth.dto.request.AuthSignUpRequestDto;
 import sparta.todoapp.global.config.security.jwt.JwtUtil;
 import sparta.todoapp.global.error.exception.ServiceException;
 import sparta.todoapp.test.ControllerTest;
