+++ src/test/java/sparta/todoapp/domain/auth/service/AuthServiceTest.java
@@ -18,8 +18,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.ActiveProfiles;
-import sparta.todoapp.domain.auth.dto.AuthLoginRequestDto;
-import sparta.todoapp.domain.auth.dto.AuthSignUpRequestDto;
+import sparta.todoapp.domain.auth.dto.request.AuthLoginRequestDto;
+import sparta.todoapp.domain.auth.dto.request.AuthSignUpRequestDto;
 import sparta.todoapp.domain.auth.entity.User;
 import sparta.todoapp.domain.auth.entity.UserRoleEnum;
 import sparta.todoapp.domain.auth.repository.UserRepository;
